% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_bstats.R
\name{abundance}
\alias{abundance}
\title{Relative abundance calculation:}
\usage{
abundance(vec, minRA = 0)
}
\arguments{
\item{vec}{Numeric vector to transform in abundances.}

\item{minRA}{Minimum RA value to filter. Defaults to 0 (no filtering).}
}
\description{
Calculates relative abundance (RA) in percentage (sum = 100) from a vector and filters by a minimum value. After filtering,
it recalculates RA to sum 100.
}
\examples{
a <- seq(1:20)
abundance(a)
# check if abundance vector sums 100:
sum(abundance(a))
}
\keyword{abundance}
\keyword{filter}
