% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_aldex2.R
\name{aldex.effect.glm}
\alias{aldex.effect.glm}
\title{Effect size calculation for GLM model:}
\usage{
aldex.effect.glm(clr, metadata, factor, levels)
}
\arguments{
\item{clr}{Data frame transformed to CLR.}

\item{metadata}{Data frame with phenotypic variables including factors.}

\item{factor}{2-length vector with factor names.}

\item{levels}{??}
}
\description{
\code{microDA} function from the \code{funs_aldex2} group (expanding \code{ALDEx2} functionality).\cr
Calculates effet size when using GLM with two factors in a similar way to \code{\link[ALDEx2]{aldex.effect}}.\cr
\code{ALDEx2} package required. THIS IS A TEST!!!
}
\keyword{ALDEx2}
\keyword{effect}
\keyword{glm}
