% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_aldex2.R
\name{aldex.ggplot}
\alias{aldex.ggplot}
\title{Differential abundance plot from ALDEx2:}
\usage{
aldex.ggplot(
  aldex.test,
  aldex.effect,
  taxonomy = NULL,
  taxrank = NULL,
  test = "welch",
  cutoff = 0.1,
  ra_thr = 1,
  layer = F,
  psize = 3
)
}
\arguments{
\item{aldex.test}{Data frame from \code{\link[ALDEx2]{aldex.ttest}} function or similar.}

\item{aldex.effect}{Output from \code{\link[ALDEx2]{aldex.effect}} function.}

\item{taxonomy}{Add taxonomy if wanted. Defaults to NULL.}

\item{taxrank}{Taxonomy rank column corresponding to feature names from aldex.effect.
Defaults to "row.names" when a taxonomy data frame is given. If \code{taxonomy = NULL}, it gets NULL value aswell.}

\item{test}{Used statistical test to make aldex.test data frame. Defaults to "welch".}

\item{cutoff}{Significance p-value threshold. Defaults to 0.1.}

\item{ra_thr}{Relative abundance threshold to consider differences as big enough. Defaults to 1.}

\item{layer}{Adds predefined standard title, subtitle and captions to plot. Defaults to FALSE.}

\item{psize}{Size of plot points. Defaults to 3.}
}
\description{
\code{microDA} function from the \code{funs_aldex2} group (expanding \code{ALDEx2} functionality).\cr
Builds a differential abundance volcano plot-like graph with ALDEx2 objects.\cr
\code{ALDEx2} package required.
}
\keyword{ALDEx2}
\keyword{abundance}
\keyword{volcano}
