% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_aldex2.R
\name{aldex.ra.out}
\alias{aldex.ra.out}
\title{Don't remember...}
\usage{
aldex.ra.out(effect.glm, raw.data, sampletax, closured = FALSE)
}
\arguments{
\item{effect.glm}{??}

\item{raw.data}{??}

\item{sampletax}{??}

\item{closured}{??}
}
\description{
\code{microDA} function from the \code{funs_aldex2} group (expanding \code{ALDEx2} functionality).\cr
Does something with RA.\cr
\code{ALDEx2} package required. THIS IS A TEST!!!
}
\keyword{ALDEx2}
\keyword{abundance}
