% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_analysis.R
\name{ALR_calc}
\alias{ALR_calc}
\title{Additive log-ratio (ALR) transformation of a compositional dataset:}
\usage{
ALR_calc(
  data,
  reference,
  zerorep = "cmultrepl",
  closure = 100,
  transpose = TRUE,
  alrweights = FALSE
)
}
\arguments{
\item{data}{Data frame to transform.}

\item{reference}{Name of the feature to be used as reference.}

\item{zerorep}{Zero replacement method. Can take \strong{"cmultrepl"} (default, bayesian imputation), \strong{0}
(no replacement) or any other number (constant replacement by specified value).}

\item{closure}{Relative abundances will be calculated to sum up to this number per sample. Defaults to 100.}

\item{transpose}{Whether the matrix should be transposed or not, since samples must be in rows and features in columns.
Defaults TRUE.}

\item{alrweights}{Custom weights for each feature. Either a vector with weights or FALSE (default, no weights).}
}
\description{
\code{microDA} function from the \code{funs_analysis} group.\cr
Includes multiple options to obtain ALR transformation of a matrix, at several steps:\cr
\enumerate{
\item Zero replacement: no replacement, cmultRepl imputation or constant value replacement.\cr
\item ALR weights: equal weights or custom.\cr
\item CLR transformation: easyCODA package.\cr
\code{easyCODA} and \code{zCompositions} packages required.\cr
}
}
\keyword{ALR}
\keyword{easyCODA}
