% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_ggplot.R
\name{CIplot_biv}
\alias{CIplot_biv}
\title{Scatterplot of groups means and ellipses:}
\usage{
CIplot_biv(
  x,
  y,
  group,
  wt = rep(1/length(x), length(x)),
  varnames = c("x", "y"),
  groupnames = sort(unique(group)),
  groupcols = rainbow(length(unique(group))),
  shownames = TRUE,
  showcentr = FALSE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  lty = 1,
  lwd = 1,
  add = FALSE,
  alpha = 0.95,
  ellipse = 0,
  shade = FALSE,
  frac = 0.01,
  cex = 1.2
)
}
\arguments{
\item{x}{The x-variable.}

\item{y}{The y-variable.}

\item{group}{The grouping variable.}

\item{wt}{Set of weights on the cases (operates when ellipse=1).}

\item{varnames}{Vector of two labels for the axes (default x and y).}

\item{groupnames}{Vector of labels for the groups (default is 1, 2, etc...).}

\item{groupcols}{Vector a colours for the groups.}

\item{shownames}{Shows group centroids as a label (default=TRUE).}

\item{showcentr}{Shows group centroids as a dot (default=FALSE).}

\item{xlim}{Possible new limits for the plot's x axis.}

\item{ylim}{Possible new limits for the plot's y axis.}

\item{lty}{Line type for the ellipse (default is 1).}

\item{lwd}{Line width for the ellipse (default is 1).}

\item{add}{If TRUE, ellipses/intervals are added to existing plot (default=FALSE).}

\item{alpha}{Ellipses confidence level (default=0.95).}

\item{ellipse}{Set ellipse<0 for regular data-covering ellipses; ellipse=0 (default) for normal-theory confidence ellipses;
ellipse=1 for bootstrap confidence ellipses; ellipse=2 for confidence error bars along ellipse axes (not implemented yet!);
ellipse=3 for normal-theory confidence error bars lined up with axes; ellipse=4 for bootstrap confidence error bars along axes.}

\item{shade}{TRUE for ellipse shading (default=FALSE).}

\item{frac}{If value = 'proportional', part defining the width of the bars at the edges of confidence intervals
(for ellipse = 3 and 4).}

\item{cex}{Character expansion factor for group names.}
}
\description{
\code{microDA} function from the \code{funs_ggplot} group (expanding \code{baseplot} functionality).\cr
Builds an x-y scatterplot of groups means and 95\% confidence ellipses classifying the data into groups.\cr
Function made by Michael Greenacre. \code{ellipse} package required.
}
\keyword{PCA}
\keyword{ellipse}
\keyword{plot}
