% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_analysis.R
\name{CLR_calc}
\alias{CLR_calc}
\title{Centered log-ratio (CLR) transformation of a compositional dataset:}
\usage{
CLR_calc(
  data,
  zerorep = "cmultrepl",
  mode = "eCODA",
  closure = 100,
  transpose = TRUE,
  clrweights = FALSE
)
}
\arguments{
\item{data}{Data frame to transform.}

\item{zerorep}{Zero replacement method. Can take \strong{"cmultrepl"} (default, bayesian imputation), \strong{"const"}
(replacement by 0.5\*DL), \strong{"unif"} (replacement by a uniform distribution between 0.1\*DL and DL), \strong{0}
(no replacement) or any other number (constant replacement by specified value).}

\item{mode}{CLR transform method. Can take \strong{"eCODA"} (default, easyCODA::CLR function) and \strong{"compositions"}
(compositions::clr function) values.}

\item{closure}{Relative abundances will be calculated to sum up to this number per sample. Defaults to 100.}

\item{transpose}{Whether the matrix should be transposed or not, since samples must be in rows and features in columns.
Defaults TRUE.}

\item{clrweights}{Custom weights for each feature. Either a vector with weights or FALSE (default, no weights).}
}
\description{
\code{microDA} function from the \code{funs_analysis} group.\cr
Includes multiple options to obtain CLR transformation of a matrix, at several steps:\cr
\enumerate{
\item Zero replacement: no replacement, cmultRepl imputation, constant value replacement or detection limit-based replacement.\cr
\item CLR weights: equal weights or custom.\cr
\item CLR transformation: easyCODA or compositions packages.\cr
\code{easyCODA}, \code{zCompositions} and \code{compositions} packages required.\cr
}
}
\keyword{CLR}
\keyword{compositions}
\keyword{easyCODA}
