% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_ggplot.R
\name{count_distrib}
\alias{count_distrib}
\title{Count distribution plot:}
\usage{
count_distrib(data)
}
\arguments{
\item{data}{Matrix with counts of sequencing reads.}
}
\description{
\code{microDA} function that classifies reads according to their counts values and generates a count distribution plot for each sample.
}
\keyword{counts}
\keyword{distribution}
\keyword{ggplot}
