% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_bstats.R
\name{countZero}
\alias{countZero}
\title{Number of zeros:}
\usage{
countZero(arg)
}
\arguments{
\item{arg}{Numeric vector.}
}
\description{
Counts the number of zeros in a vector.
}
\examples{
a <- c(1,1,0,0,1,0)
b <- seq(1:10)
countZero(a)
countZero(b)
}
\keyword{count}
\keyword{zero}
