% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_ggplot.R
\name{Ellipse_PCA}
\alias{Ellipse_PCA}
\title{PCA plot with ellipses:}
\usage{
Ellipse_PCA(
  pca,
  phenot,
  factor,
  fcolors,
  comps = c(1, 2),
  title = TRUE,
  level = NULL
)
}
\arguments{
\item{pca}{PCA object from \code{\link{prcomp}}.}

\item{phenot}{Phenotypic data frame for factor correspondence.}

\item{factor}{Name of grouping variable.}

\item{fcolors}{Vector with color codes for groups.}

\item{comps}{Length=2 vector with PCA components to represent. Defaults to c(1,2).}

\item{title}{Whether to add plot title or not. Defaults TRUE.}

\item{level}{If title = TRUE, add a word with taxonomic (or other) level done by PCA. (must be changed to generalise).}
}
\description{
\code{microDA} function from the \code{funs_ggplot} group (expanding \code{baseplot} functionality).\cr
Builds a PCA plot adding centroids and ellipses using \code{\link[microDA]{CIplot_biv}} function.
}
\keyword{PCA}
\keyword{ellipse}
\keyword{plot}
