% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_phy.R
\name{get_max_taxonomic_rank}
\alias{get_max_taxonomic_rank}
\title{Get max taxonomic rank:}
\usage{
get_max_taxonomic_rank(x, return_rank_only = FALSE)
}
\arguments{
\item{x}{Either a phyloseq object, or a data frame with columns as taxonomic ranks and rows as entries (e.g., OTUs). Columns in the data frame should be ordered from the highest level of classification (e.g., Kingdom) to the lowest level (e.g., Species), missing data are coded as NA.}

\item{return_rank_only}{Logical, if TRUE only name of the taxonomic rank will be returned}
}
\description{
\code{microDA} function from the \code{funs_phy} group (expanding \code{phyloseq} functionality).\cr
Determines the lowest level of taxonomic classification, finds the last non-NA column in the taxonomy table and returns it.\cr
Function adapted from metagMisc repository (https://github.com/vmikk/metagMisc).
}
\keyword{phyloseq}
\keyword{taxonomy}
