% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_ggplot.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Add color palette based in hue:}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of colors to include.}
}
\description{
\code{microDA} function from the \code{funs_ggplot} group (expanding \code{ggplot2} functionality).\cr
Builds a color palette based in hues from \code{\link{hcl}} function.
}
\details{
This function uses \code{\link{hcl}} specifying luminosity 65 and chroma 100 but changing the hue, which is generated
sequentially from 15 to 375 according to the number of colors desired (n).
}
\examples{
a <- gg_color_hue(10)
# Visualize colors:
colortools::pizza(a)
}
\keyword{color}
\keyword{hcl}
\keyword{hue}
