% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_ggplot.R
\name{ggrare}
\alias{ggrare}
\title{Rarefaction curves:}
\usage{
ggrare(
  physeq,
  step = 10,
  color = NULL,
  label = NULL,
  plot = T,
  parallel = F,
  se = T
)
}
\arguments{
\item{physeq}{phyloseq class object, from which abundance data are extracted.}

\item{step}{Step size for sample size in rarefaction curves.}

\item{color}{Character string. The name of the variable to map to colors in the plot. This can be a sample variable
(among the set returned by \code{sample_variables(physeq)}) or taxonomic rank (among the set returned by
\code{rank_names(physeq)}). Finally, The color scheme is chosen automatically by \code{\link[ggplot2]{ggplot}}, but it can
be modified afterward with an additional layer using \code{\link[ggplot2]{scale_colour_manual}}. Defaults NULL.}

\item{label}{Character string. The name of the variable to map to text labels on the plot. Similar to color option but for
plotting text. Defaults NULL.}

\item{plot}{Logical, whether the graphic should be plotted or not. Defaults TRUE.}

\item{parallel}{Whether rarefaction should be parallelized or not (using parallel framework).}

\item{se}{Logical, whether standard errors should be computed or not. Defaults TRUE.}
}
\description{
\code{microDA} function from the \code{funs_ggplot} group (expanding \code{ggplot2} functionality).\cr
Builds a color palette based in hues from \code{\link{hcl}} function.
}
\details{
This function is adapted from \code{vegan} \code{\link[vegan]{rarecurve}} function. It adds a ggplot for
representing rarefaction curves.
}
\keyword{ggplot}
\keyword{rarefaction}
\keyword{vegan}
