% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_ggplot.R
\name{make_sunburst}
\alias{make_sunburst}
\title{Sunburst from data:}
\usage{
make_sunburst(sunreads, sunfeat, nranks, aggregate = T, plottype = 1)
}
\arguments{
\item{sunreads}{Feature reads data frame.}

\item{sunfeat}{Feature taxonomy (or other) classification.}

\item{nranks}{Number of ranks in taxonomic (or other) classification.}

\item{aggregate}{Whether glomming reads by sunburst classification or not (in case any class is repeated). Defaults TRUE.}

\item{plottype}{Values can be 1 (\code{\link[sunburstR]{sunburst}} function) or 2 (\code{\link[sunburstR]{sund2b}} function).}
}
\description{
\code{microDA} function from the \code{funs_ggplot} group (expanding \code{ggplot2} functionality).\cr
Builds a \code{sunburstR} plot from a microbiome data frame.
}
\keyword{sunburst}
\keyword{sunburstR}
