% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_bstats.R
\name{Normtest_oneway}
\alias{Normtest_oneway}
\title{Normality and homocedasticity test:}
\usage{
Normtest_oneway(data, factor, vars)
}
\arguments{
\item{data}{Data frame with both data and factor as columns, and samples as rows.}

\item{factor}{String with the factor name.}

\item{vars}{String vector with names numeric variables to measure.}
}
\description{
Performs a normality and homocedasticity test for a dataset accounting each level of one factor. It also performs a
statistical test according to the normality obtained results. The \strong{\code{nortest}} package is required.
}
\details{
This function selects different test according to dataset properties. Being \strong{n} the number of samples and
\strong{k} the levels of the factor:\cr

Normality test: H0 = Normal distribution. With p < 0.05, H0 is rejected.\cr
\verb{   }n <= 50 -> \code{\link[nortest]{lillie.test}}\cr
\verb{   }n > 50 -> \code{\link{shapiro.test}}\cr

Variance equality test: H0 = Equal variances. With p < 0.05, H0 is rejected.\cr
\verb{   }k = 2 -> \code{\link{var.test}}\cr
\verb{   }k > 2 -> \code{\link{bartlett.test}}\cr

Statistical test:\cr
\verb{   }Normal distribution & k = 2: \code{\link{t.test}}\cr
\verb{   }Normal distribution & k > 2: \code{\link{aov}}\cr
\verb{   }Not normal distribution & k = 2: \code{\link{wilcox.test}}\cr
\verb{   }Not normal distribution & k > 2: \code{\link{kruskal.test}}
}
\keyword{homocedasticity}
\keyword{normality}
\keyword{test}
