% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_phy.R
\name{phy_filltax}
\alias{phy_filltax}
\title{Filling blank taxonomy:}
\usage{
phy_filltax(data, physeq = TRUE, unclassified = FALSE)
}
\arguments{
\item{data}{Input dataset. Either a data frame or a \code{phyloseq} object.}

\item{physeq}{Whether input is a phyloseq object or not. Defaults TRUE.}

\item{unclassified}{Whether to use 'Unclassified' as prefix for filling (TRUE) or
\code{c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus")} as prefix (FALSE). Defaults FALSE.}
}
\description{
\code{microDA} function from the \code{funs_phy} group (expanding \code{phyloseq} functionality).\cr
Fills blank or NA cells of a taxonomy table using the information of the superior level. Input data can be either a data frame
or a \code{tax_table} from a phyloseq object.
}
\details{
Time and again taxonomy tables from microbiome analysis software do not have a complete taxonomy, and one OTU may be
classified to Phylum level while other may be classified to Species level, leaving blank cells in the table. This function
fills the gaps using a prefix plus the information from the former cell.

As an example:\cr
OTU1: Kingdom: Bacteria; Phylum: NA

With \code{unclassified = FALSE}: \cr
OTU1: Kingdom: Bacteria; Phylum: Unclassified_Bacteria; Class: Unclassified_Bacteria.

With \code{unclassified = TRUE}: \cr
OTU1: Kingdom: Bacteria; Phylum: Kingdom_Bacteria; Class: Phylum_Bacteria.

Note that input data must have \strong{7 taxonomy columns}.
}
\keyword{fill}
\keyword{phyloseq}
\keyword{taxonomy}
