% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_phy.R
\name{phy_rm_na_tax}
\alias{phy_rm_na_tax}
\title{Remove NA taxonomy columns:}
\usage{
phy_rm_na_tax(physeq)
}
\arguments{
\item{physeq}{Input phyloseq object.}
}
\description{
\code{microDA} function from the \code{funs_phy} group (expanding \code{phyloseq} functionality).\cr
This function is useful after applying \code{\link[phyloseq]{tax_glom}} to a phyloseq object. It removes all the taxonomy
columns from \code{tax_table(physeq)} composed exclusively by NAs.
}
\keyword{NA}
\keyword{glom}
\keyword{phyloseq}
\keyword{taxonomy}
