% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_phy.R
\name{phy_transform}
\alias{phy_transform}
\title{Phyloseq subsetting to core:}
\usage{
phy_transform(physeq, transf = "compositional", detection, prevalence)
}
\arguments{
\item{physeq}{\code{phyloseq} input object.}

\item{transf}{Transformation to apply previous to the subsetting. See \code{\link[microbiome]{transform}} for more details.
Can also take NULL value (no transformation previous to subsetting).}

\item{detection}{Detection threshold for absence/presence. See \code{\link[microbiome]{core}} for more details.}

\item{prevalence}{Prevalence threshold (in \[ 0, 1 \]). See \code{\link[microbiome]{core}} for more details.}
}
\value{
\item{Reads}{Input phyloseq object.}

\item{RA}{Transformed initial phyloseq object. If \code{transf = NULL} it will be the same as \code{Reads} component.}

\item{Core}{Subsetted phyloseq. Transformed to compositional.}
}
\description{
\code{microDA} function from the \code{funs_phy} group (expanding \code{phyloseq} functionality).\cr
Subsets a \code{phyloseq} object given \strong{detection} and \strong{prevalence} thresholds.
}
\details{
This function uses \code{\link[microbiome]{transform}} and \code{\link[microbiome]{core}} functions to get a
compositional-transformed phyloseq including only prevalent taxa.
}
\keyword{core}
\keyword{phyloseq}
\keyword{transform}
