% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_phy.R
\name{phyloseq_to_df}
\alias{phyloseq_to_df}
\title{Phyloseq to data frame:}
\usage{
phyloseq_to_df(
  physeq,
  addtax = T,
  addtot = F,
  addmaxrank = F,
  sorting = "abundance"
)
}
\arguments{
\item{physeq}{\code{phyloseq} input object.}

\item{addtax}{Whether adding \code{tax_table(physeq)} to new data frame or not. Defaults TRUE.}

\item{addtot}{Whether adding total OTU abundance as a final column or not. Defaults FALSE.}

\item{addmaxrank}{Whether adding a max taxonomic rank column or not. Defaults FALSE.}

\item{sorting}{Reorder OTUs. Can take \strong{"abundance"} (default), \strong{"taxonomy"} or \strong{NULL} values.}
}
\description{
\code{microDA} function from the \code{funs_phy} group (expanding \code{phyloseq} functionality).\cr
Converts a \code{phyloseq} object into a \code{data frame}.\cr
Function adapted from metagMisc repository (https://github.com/vmikk/metagMisc).
}
\keyword{dataframe}
\keyword{phyloseq}
