% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_ggplot.R
\name{plotVenn}
\alias{plotVenn}
\title{Venn diagram plot:}
\usage{
plotVenn(vlist, ...)
}
\arguments{
\item{vlist}{List of character strings (max list length = 4).}

\item{...}{Other arguments from \code{VennDiagram} functions, such as \code{category} for headers of the diagram or \code{fill} for colors.}
}
\description{
\code{microDA} function from the \code{funs_ggplot} group (expanding \code{VennDiagram} functionality).\cr
Builds an expanded Venn Diagram using a list of string vectors (up to 4). Requires \code{VennDiagram} package.
}
\examples{
# Create vectors with common and uncommon elements and group them in a named list:
a <- c("one", "two", "four", "ten")
b <- c("one", "two", "five", "seven", "eight", "ten")
c <- c("one", "three", "seven", "nine", "eleven")
vlist <- list(avec = a, bvec = b, cvec = c)
# Plot Venn diagram:
plotVenn(vlist, category = names(vlist), fill = c("red", "blue", "green"))
}
\keyword{PCA}
\keyword{ellipse}
\keyword{plot}
