% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_bstats.R
\name{relAbundance}
\alias{relAbundance}
\title{relAbundance:}
\usage{
relAbundance(abundance, taxonomy_tree, level)
}
\arguments{
\item{abundance}{???}

\item{taxonomy_tree}{???}

\item{level}{???}
}
\description{
Does something related to \code{abundance} function.
}
\keyword{abundance}
