% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_analysis.R
\name{REPLzeros}
\alias{REPLzeros}
\title{Detection limit-based zero replacement on compositional data:}
\usage{
REPLzeros(comp, method)
}
\arguments{
\item{comp}{Input composition (i.e., one sample with compositional features).}

\item{method}{Zero replacement method. Can take \strong{"const"} (replacement by 0.5\*DL) and \strong{"unif"} (replacement
by a uniform distribution between 0.1\*DL and DL) values.}
}
\description{
\code{microDA} function from the \code{funs_analysis} group that applies detection limit (DL)-based zero replacement
methods on individual compositions.\cr
}
\details{
The two methods computed here are described in Lubbe et al (2021) (\url{https://doi.org/10.1016/j.chemolab.2021.104248}).
}
\keyword{CLR}
\keyword{compositions}
\keyword{easyCODA}
