% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_bstats.R
\name{sem}
\alias{sem}
\title{Standard Error of the Mean}
\usage{
sem(x)
}
\arguments{
\item{x}{Numeric data vector.}
}
\description{
Calculates standard error of the mean (sem) in one vector.
}
\examples{
a <- seq(1:10)
sem(a)
}
\keyword{sem}
