% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_deseq2.R
\name{sigtabfun}
\alias{sigtabfun}
\title{DESeq2 expanded results table:}
\usage{
sigtabfun(
  dsq,
  contrast,
  siglevel,
  lfcCutoff,
  phyloseq,
  taxorder,
  abund.grps = contrast[1],
  RA = F
)
}
\arguments{
\item{dsq}{Deseq object.}

\item{contrast}{String vector with 3 elements: Name of the factor to separate data, group-1 and group-2.}

\item{siglevel}{q-value threshold.}

\item{lfcCutoff}{Fold-change threshold.}

\item{phyloseq}{Phyloseq object used to build the RA table and the Deseq object.}

\item{taxorder}{Name of taxonomic level to reorder rows from output table.}

\item{abund.grps}{Groups used to calculate average abundances. Defaults to \code{contrast[1]}.}

\item{RA}{Logical. Whether to show abundances in terms of relative abundance (TRUE) or deseq-normalized counts (FALSE). Defaults TRUE.}
}
\description{
\code{microDA} function from the \code{deseq2} group (expanding \code{DESeq2} functionality).\cr
Builds an expanded DESeq2 results table filtering by q-value and fold-change significance thresholds, and adding average
group abundances. Requires \code{DESeq2} package.
}
\keyword{DESeq2}
\keyword{phyloseq}
