% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microDA_funs_ggplot.R
\name{ggpca}
\alias{ggpca}
\title{Principal Component Analysis plot with centroids:}
\usage{
ggpca(pca, pdata, fit = NULL, group, axes = c(1, 2))
}
\arguments{
\item{pca}{PCA object from \code{\link{prcomp}}.}

\item{pdata}{Phenotypic data frame for factor correspondence.}

\item{fit}{Whether to add linear fitting (using \code{ggordiplots} \code{\link[ggordiplots]{gg_envfit}} function) of
variable to PCA plot or not. Defaults to NULL (no fitting).}

\item{group}{Factor name for centroid grouping.}

\item{axes}{Length=2 vector with PCA axes to represent. Defaults to c(1,2).}

\item{centr}{Define if group centroids might be represented. Defaults FALSE.}
}
\description{
\code{microDA} function from the \code{funs_ggplot} group (expanding \code{ggplot2} functionality).\cr
Builds a PCA plot adding centroids according to a user-defined factor.
}
\keyword{PCA}
\keyword{envfit}
\keyword{ggplot}
